+++ src/ScottPlot4/changelog.md
@@ -8,6 +8,7 @@ _not yet published on NuGet..._
 * HLine and VLine: Line (but not position label) is hidden if `LineWidth` is `0` (#2085) _Thanks @A1145681_
 * Controls: The cursor now reverts to `Configuration.DefaultCursor` after moving off draggable objects (#2091) _Thanks @kurupt44_
 * Snapping: SnapNearest classes now expose `SnapIndex()` (#2099) _Thanks @BambOoxX_
+* Background: Added optional arguments to `Style()` lets users place a custom background image behind their plot (#2016) _Thanks @apaaris_
 
 ## ScottPlot 4.1.57
 _Published on [NuGet](https://www.nuget.org/profiles/ScottPlot) on 2022-08-18_